Add Open in new tab option based on Advanced URI plugin
Add the Open in new tab by default option and a result in context menu when option is not activated, Both need Advanced URI plugin to work.

Disable the option and show a tooltip of why, when the Advanced URI plugin is not detected in vault/globally.

Reformat GlobalVaultSetting.cs for clarity.